Fix nightly_utils import: move to sglang.test module
Move nightly_utils.py from test/nightly/ to python/sglang/test/
so it can be imported as sglang.test.nightly_utils by the migrated
perf tests in test/registered/perf/.
