+++ test/nightly/performance_test_runner.py
@@ -1,7 +1,7 @@
 from dataclasses import dataclass, field
 from typing import List, Optional, Tuple
 
-from nightly_utils import NightlyBenchmarkRunner
+from sglang.test.nightly_utils import NightlyBenchmarkRunner
 
 from sglang.test.nightly_bench_utils import BenchmarkResult
 from sglang.test.test_utils import DEFAULT_URL_FOR_TEST, ModelLaunchSettings
