+++ test/registered/perf/test_gpt_oss_4gpu_perf.py
@@ -1,8 +1,7 @@
 import unittest
 
-from nightly_utils import NightlyBenchmarkRunner
-
 from sglang.test.ci.ci_register import register_cuda_ci
+from sglang.test.nightly_utils import NightlyBenchmarkRunner
 from sglang.test.test_utils import DEFAULT_URL_FOR_TEST
 
 register_cuda_ci(est_time=600, suite="nightly-4-gpu-b200", nightly=True)
