+++ test/registered/perf/test_text_models_perf.py
@@ -1,8 +1,7 @@
 import unittest
 
-from nightly_utils import NightlyBenchmarkRunner
-
 from sglang.test.ci.ci_register import register_cuda_ci
+from sglang.test.nightly_utils import NightlyBenchmarkRunner
 from sglang.test.test_utils import (
     DEFAULT_URL_FOR_TEST,
     ModelLaunchSettings,
